requirements.txt
pyspark==3.5.1
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
kafka-python==2.0.2
pyarrow==16.1.0
streamlit==1.38.0
pymongo==4.8.0
python-dotenv==1.0.1
matplotlib==3.9.1
seaborn==0.13.2
ujson==5.10.0


